Crypto Icons
Who needs content when you have icons? I've added a whopping 1932 crypto icons to the project, sourced from the trusty ol' cryptoicons website. Because let's be real, the only thing that matters in a crypto app is how many icons you have. Users will be so dazzled by the sheer number of icons that they won't even notice the lack of actual functionality. So go ahead and bask in the glory of these tiny, pixelated masterpieces. Who needs a working product when you have pretty icons?